refactor: 램덤 숫자를 생성하는 방식을 인터페이스를 구현하여 생성하는 방식으로 변경
테스트 코드 작성을 쉽게 만들어주고, 추후의 변화에도 유동적으로 변경할 수 있도록 구조 변경
