+++ src/main/java/racingcar/domain/car/Cars.java
@@ -1,7 +1,5 @@
 package racingcar.domain.car;
 
-import racingcar.domain.car.util.RandomNumberGenerator;
-
 import java.util.*;
 
 public class Cars {
@@ -21,11 +19,10 @@ public class Cars {
         }
     }
 
-    public Cars play(RandomNumberGenerator randomNumberGenerator) {
+    public Cars play(NumberGenerator numberGenerator) {
         List<Car> cars = new ArrayList<>();
-
         for (Car car : this.cars) {
-            cars.add(car.move(randomNumberGenerator.generate()));
+            cars.add(car.move(numberGenerator.generate()));
         }
 
         return new Cars(cars);
