+++ src/main/java/racingcar/domain/car/NumberGenerator.java
@@ -1,0 +1,7 @@
+package racingcar.domain.car;
+
+@FunctionalInterface
+public interface NumberGenerator {
+
+    int generate();
+}
