+++ src/main/java/racingcar/domain/car/RandomNumberGenerator.java
@@ -1,11 +1,12 @@
-package racingcar.domain.car.util;
+package racingcar.domain.car;
 
 import java.util.Random;
 
-public class RandomNumberGenerator {
+public class RandomNumberGenerator implements NumberGenerator {
     private static final Random RANDOM = new Random();
     private static final int END_EXCLUSIVE_NUMBER = 10;
 
+    @Override
     public int generate() {
         return RANDOM.nextInt(END_EXCLUSIVE_NUMBER);
     }
