+++ src/test/java/racingcar/domain/car/CarsTest.java
@@ -2,7 +2,6 @@ package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.util.RandomNumberGenerator;
 
 import java.util.Arrays;
 import java.util.List;
@@ -59,12 +58,12 @@ class CarsTest {
         Cars cars = new Cars(carGroup);
 
         //when
-        Cars movedCars = cars.play(new RandomNumberGenerator());
+        Cars movedCars = cars.play(() -> 4);
 
         //then
         List<Car> carsValue = movedCars.getCars();
-        assertThat(carsValue.get(0).getPosition()).isGreaterThanOrEqualTo(0).isLessThanOrEqualTo(1);
-        assertThat(carsValue.get(1).getPosition()).isGreaterThanOrEqualTo(0).isLessThanOrEqualTo(1);
-        assertThat(carsValue.get(2).getPosition()).isGreaterThanOrEqualTo(0).isLessThanOrEqualTo(1);
+        assertThat(carsValue.get(0).getPosition()).isEqualTo(1);
+        assertThat(carsValue.get(1).getPosition()).isEqualTo(1);
+        assertThat(carsValue.get(2).getPosition()).isEqualTo(1);
     }
 }
